Safari-609.1.11.1/LayoutTests/js/dom/navigator-webdriver.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Check that navigator.webdriver has the right attributes.");

if (window.testRunner)
    testRunner.dumpAsText();

shouldBeTrue("'webdriver' in window.navigator");
shouldBeFalse("navigator.webdriver");

var descriptor = Object.getOwnPropertyDescriptor(navigator, "webdriver") ||
    Object.getOwnPropertyDescriptor(Object.getPrototypeOf(navigator), "webdriver");

shouldBeTrue("descriptor.configurable");
shouldBeTrue("descriptor.enumerable");
shouldBeUndefined("descriptor.set");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
