LayoutTests/js/dom/navigator-webdriver-expected.txt

Check that navigator.webdriver has the right attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webdriver' in window.navigator is true
PASS navigator.webdriver is false
PASS descriptor.configurable is true
PASS descriptor.enumerable is true
PASS descriptor.set is undefined.
PASS successfullyParsed is true

TEST COMPLETE
